refactor(core): rename JSON reader module and unify Bookmarks reasons
Address final-round review nits without changing Analyzer output.

- Rename preferences-json.ts -> working-copy-json.ts (and its test) and
  update the module doc: after the earlier extraction it is the generic
  verified-JSON Working Copy reader used by Preferences and Bookmarks
  alike, so the preferences-specific name and doc were stale and would
  misdirect the next JSON artifact.
- Label-prefix the Bookmarks unsupported-shape reason so a primary
  reports `bookmarks_unsupported_shape` and a backup reports
  `bookmarks_backup_unsupported_shape`, consistent with the malformed and
  entry-state branches. Now every failure reason names its source file,
  so primary and backup stay distinguishable by reason string alone.

Pure rename + reason-string consistency: no successful-analysis output
changes. pnpm verify green (112 tests).
